Remove duplicate auth context
