--- context/AuthContext.tsx
@@ -1,49 +1,0 @@
-'use client'
-
-import { createContext, useContext, useEffect, useState, ReactNode } from 'react'
-import { onAuthStateChanged, User } from 'firebase/auth'
-import { doc, getDoc } from 'firebase/firestore'
-import { auth } from '@/lib/firebase/init'
-import { db } from '@/lib/firebase'
-
-interface AuthContextType {
-  user: User | null
-  userData: any
-  loading: boolean
-}
-
-export const AuthContext = createContext<AuthContextType>({
-  user: null,
-  userData: null,
-  loading: true
-})
-
-export const AuthProvider = ({ children }: { children: ReactNode }) => {
-  const [user, setUser] = useState<User | null>(null)
-  const [userData, setUserData] = useState<any>(null)
-  const [loading, setLoading] = useState(true)
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, async (currentUser) => {
-      setUser(currentUser)
-      if (currentUser) {
-        const ref = doc(db, 'users', currentUser.uid)
-        const snap = await getDoc(ref)
-        setUserData(snap.exists() ? snap.data() : null)
-      } else {
-        setUserData(null)
-      }
-      setLoading(false)
-    })
-
-    return () => unsubscribe()
-  }, [])
-
-  return (
-    <AuthContext.Provider value={{ user, userData, loading }}>
-      {children}
-    </AuthContext.Provider>
-  )
-}
-
-export const useAuth = () => useContext(AuthContext)
