+++ src/context/AuthContext.tsx
@@ -1,21 +1,48 @@
 'use client'
 
-import { createContext, useContext, useEffect, useState } from 'react'
+import { createContext, useContext, useEffect, useState, ReactNode } from 'react'
 import { onAuthStateChanged, User } from 'firebase/auth'
-import { auth } from '@/lib/firebase'
+import { doc, getDoc } from 'firebase/firestore'
+import { auth, db } from '@/lib/firebase'
 
-// ✅ Explicitly export AuthContext
-export const AuthContext = createContext<{ user: User | null }>({ user: null })
+interface AuthContextType {
+  user: User | null
+  userData: any
+  loading: boolean
+}
 
-export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
+export const AuthContext = createContext<AuthContextType>({
+  user: null,
+  userData: null,
+  loading: true,
+})
+
+export const AuthProvider = ({ children }: { children: ReactNode }) => {
   const [user, setUser] = useState<User | null>(null)
+  const [userData, setUserData] = useState<any>(null)
+  const [loading, setLoading] = useState(true)
 
   useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (user) => {
-      setUser(user)
+    const unsubscribe = onAuthStateChanged(auth, async (currentUser) => {
+      setUser(currentUser)
+      if (currentUser) {
+        const ref = doc(db, 'users', currentUser.uid)
+        const snap = await getDoc(ref)
+        setUserData(snap.exists() ? snap.data() : null)
+      } else {
+        setUserData(null)
+      }
+      setLoading(false)
     })
+
     return () => unsubscribe()
   }, [])
 
-  return <AuthContext.Provider value={{ user }}>{children}</AuthContext.Provider>
+  return (
+    <AuthContext.Provider value={{ user, userData, loading }}>
+      {children}
+    </AuthContext.Provider>
+  )
 }
+
+export const useAuth = () => useContext(AuthContext)
